refactor(order-book): move `compute_impact_price` from perps to order-book
Impact price (VWAP for a hypothetical fill of a given notional size) is a
generic order-book concept, not specific to perpetual futures. Move the
function and its unit tests to `dango-order-book` so spot, options, and
prediction-market books can reuse it.

The function was dead code in the perps crate — the funding rate mechanism
now uses the vault-skew-based premium model exclusively.

---------
